refactor: update src folder structure
