--- src/api/index.tsx
@@ -1,3 +1,0 @@
-export * from './common';
-export * from './posts';
-export * from './types';
