--- src/api/posts/index.ts
@@ -1,4 +1,0 @@
-export * from './types';
-export * from './use-add-post';
-export * from './use-post';
-export * from './use-posts';
