--- src/api/posts/types.ts
@@ -1,6 +1,0 @@
-export type Post = {
-  userId: number;
-  id: number;
-  title: string;
-  body: string;
-};
