--- src/api/posts/use-post.ts
@@ -1,17 +1,0 @@
-import type { AxiosError } from 'axios';
-import type { Post } from './types';
-
-import { createQuery } from 'react-query-kit';
-import { client } from '../common';
-
-type Variables = { id: string };
-type Response = Post;
-
-export const usePost = createQuery<Response, Variables, AxiosError>({
-  queryKey: ['posts'],
-  fetcher: (variables) => {
-    return client
-      .get(`posts/${variables.id}`)
-      .then(response => response.data);
-  },
-});
