--- src/api/types.ts
@@ -1,6 +1,0 @@
-export type PaginateQuery<T> = {
-  results: T[];
-  count: number;
-  next: string | null;
-  previous: string | null;
-};
