+++ src/app/(app)/index.tsx
@@ -1,34 +1,1 @@
-import type { Post } from '@/api';
-import { FlashList } from '@shopify/flash-list';
-
-import * as React from 'react';
-import { usePosts } from '@/api';
-import { Card } from '@/components/card';
-import { EmptyList, FocusAwareStatusBar, Text, View } from '@/components/ui';
-
-export default function Feed() {
-  const { data, isPending, isError } = usePosts();
-  const renderItem = React.useCallback(
-    ({ item }: { item: Post }) => <Card {...item} />,
-    [],
-  );
-
-  if (isError) {
-    return (
-      <View>
-        <Text> Error Loading data </Text>
-      </View>
-    );
-  }
-  return (
-    <View className="flex-1">
-      <FocusAwareStatusBar />
-      <FlashList
-        data={data}
-        renderItem={renderItem}
-        keyExtractor={(_, index) => `item-${index}`}
-        ListEmptyComponent={<EmptyList isLoading={isPending} />}
-      />
-    </View>
-  );
-}
+export { FeedScreen as default } from '@/features/feed/feed-screen';
