+++ src/app/(app)/style.tsx
@@ -1,23 +1,1 @@
-import * as React from 'react';
-
-import { Buttons } from '@/components/buttons';
-import { Colors } from '@/components/colors';
-import { Inputs } from '@/components/inputs';
-import { Typography } from '@/components/typography';
-import { FocusAwareStatusBar, SafeAreaView, ScrollView } from '@/components/ui';
-
-export default function Style() {
-  return (
-    <>
-      <FocusAwareStatusBar />
-      <ScrollView className="px-4">
-        <SafeAreaView className="flex-1">
-          <Typography />
-          <Colors />
-          <Buttons />
-          <Inputs />
-        </SafeAreaView>
-      </ScrollView>
-    </>
-  );
-}
+export { StyleScreen as default } from '@/features/style-demo/style-screen';
