+++ src/app/login.tsx
@@ -1,24 +1,1 @@
-import type { LoginFormProps } from '@/components/login-form';
-import { useRouter } from 'expo-router';
-
-import * as React from 'react';
-import { LoginForm } from '@/components/login-form';
-import { FocusAwareStatusBar } from '@/components/ui';
-import { useAuth } from '@/lib';
-
-export default function Login() {
-  const router = useRouter();
-  const signIn = useAuth.use.signIn();
-
-  const onSubmit: LoginFormProps['onSubmit'] = (data) => {
-    console.log(data);
-    signIn({ access: 'access-token', refresh: 'refresh-token' });
-    router.push('/');
-  };
-  return (
-    <>
-      <FocusAwareStatusBar />
-      <LoginForm onSubmit={onSubmit} />
-    </>
-  );
-}
+export { LoginScreen as default } from '@/features/auth/login-screen';
