+++ src/features/feed/post-detail-screen.tsx
@@ -1,0 +1,49 @@
+import { Stack, useLocalSearchParams } from 'expo-router';
+import * as React from 'react';
+
+import {
+  ActivityIndicator,
+  FocusAwareStatusBar,
+  Text,
+  View,
+} from '@/components/ui';
+import { usePost } from './api';
+
+export function PostDetailScreen() {
+  const local = useLocalSearchParams<{ id: string }>();
+
+  const { data, isPending, isError } = usePost({
+    variables: { id: local.id },
+  });
+
+  if (isPending) {
+    return (
+      <View className="flex-1 justify-center p-3">
+        <Stack.Screen options={{ title: 'Post', headerBackTitle: 'Feed' }} />
+        <FocusAwareStatusBar />
+        <ActivityIndicator />
+      </View>
+    );
+  }
+  if (isError) {
+    return (
+      <View className="flex-1 justify-center p-3">
+        <Stack.Screen options={{ title: 'Post', headerBackTitle: 'Feed' }} />
+        <FocusAwareStatusBar />
+        <Text className="text-center">Error loading post</Text>
+      </View>
+    );
+  }
+
+  return (
+    <View className="flex-1 p-3">
+      <Stack.Screen options={{ title: 'Post', headerBackTitle: 'Feed' }} />
+      <FocusAwareStatusBar />
+      <Text className="text-xl">{data.title}</Text>
+      <Text>
+        {data.body}
+        {' '}
+      </Text>
+    </View>
+  );
+}
