+++ src/features/onboarding/components/cover.tsx
