+++ src/lib/api/index.ts
@@ -1,3 +1,3 @@
-export * from './api-provider';
 export * from './client';
+export * from './provider';
 export * from './utils';
