--- src/lib/index.tsx
@@ -1,4 +1,0 @@
-export * from './auth';
-export * from './hooks';
-export * from './i18n';
-export * from './utils';
